requirements.txt
aiogram==2.25.2
aiohttp
aiosqlite
aiofiles
openai-whisper
torch